Add files via upload
This shows everything collected from the event viewer when some super duper top notch spy decided to use "After Midnight" on my laptop... Or whatever the hell it was. Pretty sure it was a sophisticated attack that was meant to piss me off... but instead, gave me an opportunity to study the trajectory of the attack. Good stuff.